+++ docs/release-notes.md
@@ -92,6 +92,8 @@ and
 
 - An `OutOfMemoryError` will go through to the top. Fix for
   [bug report from jqwik-spring](https://github.com/jlink/jqwik-spring/issues/5). 
+  
+- `Arbitraries.of(listOfValues)` would break when list does not allow null values. 
 
 ## 1.2.x